refactor(web/engine): moves DOM-based OutputTargets to own namespace
fix(web): more tweaks
