+++ web/source/dom/targets/readme.md
@@ -1,0 +1,3 @@
+Please keep any code in this folder / namespace as free as possible from dependencies on other parts of KMW.  Some of our unit tests wish to run against these types without requiring KMW to be active.
+
+Note that with a little work, we _could_ completely spin this into its own separate module - this could be useful for development and testing purposes... at least, save for the TouchAlias type.
