+++ web/source/dom/targets/touchAlias.ts
@@ -1,9 +1,9 @@
 // Defines the TouchAliasElement merged type.
-/// <reference path="touchAliasElement.ts" />
+/// <reference path="../touchAliasElement.ts" />
 // Defines the OutputTarget interface.
-/// <reference path="../text/outputTarget.ts" />
+/// <reference path="../../text/outputTarget.ts" />
 
-namespace com.keyman.dom {
+namespace com.keyman.dom.targets {
   export class TouchAlias extends text.OutputTarget {
     root: TouchAliasElement;
